requirements.txt
joblib
plotly
pandas
kaleido
matplotlib
jupyter
numpy
scipy
geos
hydra-core
gradio
reverse_geocoder
scikit-learn
pre-commit
wandb
transformers
accelerate
pytorch-lightning
peft
